Move viewmodels out of widget pages
