+++ example/lib/modules/telemetry_lab_page.dart
@@ -2,6 +2,7 @@ import 'package:current/current.dart';
 import 'package:flutter/material.dart';
 
 import '../mission_control_theme.dart';
+import '../view_models/telemetry_lab_view_model.dart';
 
 class TelemetryLabPage extends CurrentWidget<TelemetryLabViewModel> {
   const TelemetryLabPage({super.key, required super.viewModel});
@@ -185,86 +186,6 @@ class _TelemetryLabPageState
     final hour = value.hour.toString().padLeft(2, '0');
     final minute = value.minute.toString().padLeft(2, '0');
     return '${value.year}-$month-$day  $hour:$minute UTC';
-  }
-}
-
-class TelemetryLabViewModel extends CurrentViewModel {
-  TelemetryLabViewModel();
-
-  final missionName = CurrentProperty.string(
-    initialValue: 'Odyssey Relay',
-    propertyName: 'missionName',
-  );
-  final crewCount = CurrentProperty.integer(
-    initialValue: 4,
-    propertyName: 'crewCount',
-  );
-  final thrust = CurrentProperty.doubleProp(
-    initialValue: 72.4,
-    propertyName: 'thrust',
-  );
-  final autopilotArmed = CurrentProperty.boolean(
-    initialValue: true,
-    propertyName: 'autopilotArmed',
-  );
-  final nextBurn = CurrentProperty.dateTime(
-    initialValue: DateTime.utc(2042, 7, 16, 14, 45),
-    propertyName: 'nextBurn',
-  );
-
-  static const _callsigns = [
-    'Odyssey Relay',
-    'Atlas Horizon',
-    'Nova Vector',
-    'Perihelion One',
-  ];
-
-  @override
-  Iterable<CurrentProperty> get currentProps => [
-        missionName,
-        crewCount,
-        thrust,
-        autopilotArmed,
-        nextBurn,
-      ];
-
-  void rotateCallsign() {
-    final currentIndex = _callsigns.indexOf(missionName.value);
-    final nextIndex =
-        currentIndex == -1 ? 0 : (currentIndex + 1) % _callsigns.length;
-    missionName.value = _callsigns[nextIndex];
-  }
-
-  void cycleCrewCount() {
-    crewCount.value = crewCount.value >= 7 ? 3 : crewCount.value + 1;
-  }
-
-  void boostThrust() {
-    thrust.value = ((thrust.value + 4.5).clamp(10.0, 100.0) as num).toDouble();
-  }
-
-  void toggleAutopilot() {
-    autopilotArmed.value = !autopilotArmed.value;
-  }
-
-  void delayBurn() {
-    nextBurn.value = nextBurn.value.add(const Duration(minutes: 90));
-  }
-
-  void simulateOrbitalBoost() {
-    setMultiple([
-      {missionName: 'Astra-Vector'},
-      {crewCount: 6},
-      {thrust: 88.2},
-      {autopilotArmed: true},
-      {nextBurn: DateTime.utc(2042, 7, 17, 6, 30)},
-    ]);
-  }
-
-  void commitBaseline() {
-    for (final property in currentProps) {
-      property.setOriginalValueToCurrent();
-    }
   }
 }
